requirements.txt
streamlit
openai==0.28
pinecone-client
streamlit
python-dotenv
